% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_N_to_catVar.R
\name{add_N_to_catVar}
\alias{add_N_to_catVar}
\title{Add sample size to the value}
\usage{
add_N_to_catVar(x)
}
\arguments{
\item{x}{A vector, of any type}
}
\value{
A \code{factor} with the label and how many times that label appeared in \code{x}
}
\description{
Add sample size to the value
}
\examples{

data(mtcars)

mtcars$cyl_n <- add_N_to_catVar(mtcars$cyl)

# you can also pipe it with tidyverse
mtcars \%>\%
mutate(cyl_n = add_N_to_catVar(cyl))

}
