% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis_to_table.R
\name{adonis_to_table}
\alias{adonis_to_table}
\title{Quick wrapper to format PERMANOVA results}
\usage{
adonis_to_table(adonis.out, new_rownames = NA)
}
\arguments{
\item{adonis.out}{A \code{adonis2} output object as is}

\item{new_rownames}{A \code{character} vector containing prettier variable
names, which will be passed with set_rownames}
}
\value{
A \code{data.frame} Ready to be passed to kable/DT/flextable
}
\description{
Quick wrapper to format PERMANOVA results
}
\examples{
library(gautils)
data(enterotype)
distMat <-  phyloseq::distance(microbiome::transform(enterotype, "compositional"), "bray")

adonisObj <- adonis2(distMat ~ SeqTech + Enterotype, by = "margin", permutations = 2000, data = microbiome::meta(enterotype), na.action = na.omit)

adonis_as_table <- adonis_to_table(adonisObj)

kableExtra::kbl(adonis_as_table, digits = 2) \%>\% 
kableExtra::kable_styling()
}
