% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_integer_into_integer.R
\name{bin_integer_into_integer}
\alias{bin_integer_into_integer}
\title{Bin an integer/numeric variable}
\usage{
bin_integer_into_integer(x, multiples, include.lowest = FALSE)
}
\arguments{
\item{x}{A `numeric` vector}

\item{multiples}{An `integer` or `numeric` vector to bin `x` by. For example, \code{multiples = 2} will return all values approximated to \code{0,2,4,6, ...}}

\item{include.lowest}{Do you want to include upper or lower values in the bin? eg: Take `multiples = 5`, the ranges will be [0,5), [5,10), ..., if FALSE, (0,5], (5,10], ...}
}
\value{
A \code{integer} vector
}
\description{
This function bins decimal numbers to their lower or upper values.
This is particularly useful for semi-quantitative modelling, in case there is a preference for a set of values.
}
\examples{
data(mtcars)
bin_integer_into_integer(mtcars$mpg, multiples = 10, include_highest= TRUE)
bin_integer_into_integer(mtcars$mpg, multiples = 5, include_highest= FALSE)

}
