% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corncob_results_into_list.R
\name{corncob_results_into_list}
\alias{corncob_results_into_list}
\title{Extract corncob results}
\usage{
corncob_results_into_list(res, class = "factor")
}
\arguments{
\item{res}{A `corncob differentialTest` result object. \\textbf{NB: only analyses with `phyloseq`objects as input are supported for now}.}

\item{class}{}

\item{trait}{A `character` vector with the metadata column name to extract the results from.}
}
\value{
A `list` of `data.frame`s with all contrasts applicable against the reference. NB: other pairwise contrasts are not made
}
\description{
An easy wrangling function to extract all results from a corncob differential abundance experiment. NB: so far two things are left to test:
(1) the order of the formulas might be important, and the last term is what this function will extract. (2) Extraction of results for regression against a continuous variable is not available yet
}
\examples{
library(corncob)
# phyloseq example
data(soil_phylum_small)
da_analysis <- differentialTest(formula = ~ DayAmdmt,
                               phi.formula = ~ DayAmdmt,
                               formula_null = ~ 1,
                              phi.formula_null = ~ DayAmdmt,
                               test = "Wald", boot = FALSE,
                               data = soil_phylum_small,
                               fdr_cutoff = 0.05)

da_analysis_wrangled <- corncob_results_into_list(da_analysis, class = "factor")

}
