% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declutter.envir.R
\name{declutter.envir}
\alias{declutter.envir}
\title{Delete temporary or test objects in `ls()`}
\usage{
declutter.envir(extra_rules = "test", verbose = FALSE)
}
\arguments{
\item{extra_rules}{A \code{character} that defines extra patterns to delete (eg: extra_rules = c("temp","delete")). NB: RegEx are also allowed}

\item{verbose}{A \code{logical} whether you want the function to print the objects it removed}
}
\description{
This function deletes all items (object, function, ...) containing either 'tmp' or 'test' in the name. Also, if the name is only 1 letter long, it will be deleted
}
\examples{

papaya <- "good"
first_attempt <- "eat"
second_attempt <- "digest"

declutter_Envir(extra_rules = c("papaya", "attempt"))

ls()
# you should only see digest
}
