% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_results_into_list.R
\name{deseq_results_into_list}
\alias{deseq_results_into_list}
\title{From DESeq2 results to a list of data.frames}
\usage{
deseq_results_into_list(
  deseq_results,
  trait,
  physeq_obj = NULL,
  taxon_lvls_added = "all",
  sort_p.val = TRUE
)
}
\arguments{
\item{deseq_results}{a \code{DESeq2} results object}

\item{trait}{A \code{character} with the variable name to compare contrasts within}

\item{physeq_obj}{the phyloseq object you used, if any}

\item{taxon_lvls_added}{A \code{character} with the taxonomic levels to add to each results \code{data.frame}}

\item{sort_p.val}{a \code{logical} saying whether you want the `padj` column sorted or not}
}
\value{
A \code{list} of pairwise contrasts based on the factor levels of the \code{trait} variable
}
\description{
This function only works with categorical contrasts
}
\examples{

library(phyloseq)
library(DESeq2)
filepath <- system.file("extdata", "study_1457_split_library_seqs_and_mapping.zip", package="phyloseq")
kostic <- microbio_me_qiime(filepath)

kostic = subset_samples(kostic, DIAGNOSIS != "None")

diagdds = phyloseq_to_deseq2(kostic, ~ OSH_DIAGNOSIS)

# calculate geometric means prior to estimate size factors

 gMeans = apply(counts(diagdds), 1, function(x, na.rm=TRUE) exp(sum(log(x[x > 0]), na.rm=na.rm) / length(x)))

# do the glm modelling

diagdds2 = estimateSizeFactors(diagdds, geoMeans = gMeans)
diagdds = DESeq(diagdds2, fitType="local")

res_list_with_Phyloseq <- deseq_results_into_list(deseq_results = diagdds,
                        trait = "OSH_DIAGNOSIS",
                        physeq_obj = kostic,
                        taxon_lvls_added = c("Phylum", "Class", "Genus")
                        )

res_list_withOut_Phyloseq <- deseq_results_into_list(deseq_results = diagdds,
                        trait = "OSH_DIAGNOSIS"
                        )
}
