% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.pval2.R
\name{format.pval2}
\alias{format.pval2}
\title{Format P-values for tables}
\usage{
\method{format}{pval2}(p, threshold = 0.001, scientific = TRUE)
}
\arguments{
\item{p}{A \code{numeric} vector of p-values to convert}

\item{threshold}{A \code{numeric} threshold under which to convert p-values
with scientific notation}

\item{scientific}{A \code{logical}, whether to use scientific notation. 
(default = `TRUE`). NOTE: `scientific = FALSE` is basically meaningless}
}
\value{
A \code{character} vector with transformed p-values
}
\description{
Easy wrapper around format.pval to make readability of p-values better. It 
converts only p-values below 0.00
}
\examples{

pvalues <- c(runif(5, 0, 0.001), runif(5, 0, 0.3), runif(5, 0, 1))

format.pval2(pvalues)

format.pval2(pvalues, threshold = 0.05)

format.pval2(pvalues, threshold = 0.05, scientific = FALSE)

}
