% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEMMA_results.R
\name{get_GEMMA_results}
\alias{get_GEMMA_results}
\title{Get Gemma heritability results}
\usage{
get_GEMMA_results(base_dir, prefix = "x__")
}
\arguments{
\item{base_dir}{\code{character} with the GEMMA base directory}

\item{prefix}{\code{chracter} specifying the trait prefix used to make trait names easily parsable. The function strips those prefixces out}
}
\value{
A `data.frame`
}
\description{
From a directory populated with GEMMA associations, this will read the chip heritability estimates
}
