% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SOLAR_results.R
\name{get_SOLAR_results}
\alias{get_SOLAR_results}
\title{Get SOLAR results}
\usage{
get_SOLAR_results(out_dir, prefix = "x__")
}
\arguments{
\item{out_dir}{A \code{character} specifying the base directory of the SOLAR analysis}
}
\value{
A `data.frame` with estimates, P-values and so on
}
\description{
This is a quick extractor of a SOLAR run results
}
