% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_target_directory.R
\name{open_target_directory}
\alias{open_target_directory}
\title{Open file/plot directory}
\usage{
open_target_directory(element_path)
}
\arguments{
\item{element_path}{The path, written unix-style. it has to be absolute, either from the root, or from the std. home ("~/")}
}
\value{
A window opened on your file browsing element ("Finder" or "Explorer")
}
\description{
Given a PATH, either ending with "/" or a file name, this function opens a window to show you that directory.
}
\examples{
# in Unix and Unix-like
open_target_directory("/home/user/Documents/my_superImportantDoc.txt")
# in Windows
open_target_directory("C:/Users/user/Documents/my_superImportantDoc.txt")


}
