% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_alpha_div_into_phy_metadata.R
\name{phy_alpha_div_into_phy_metadata}
\alias{phy_alpha_div_into_phy_metadata}
\title{Add alpha diversities into phyloseq metadata}
\usage{
phy_alpha_div_into_phy_metadata(
  physeq,
  raref_depth = NULL,
  measures,
  verbose = FALSE
)
}
\arguments{
\item{physeq}{a `phyloseq` object}

\item{raref_depth}{the number of minimum reads to rarefy by}

\item{measures}{the alpha diversity measures, as implemented by the `phyloseq::estimate_richness` function}

\item{verbose}{if `TRUE`, will tell you what it is doing, in particular in the rarefaction step}
}
\value{
a `phyloseq` object with the alpha diversity estimates added to the metadata
}
\description{
This function takes a phyloseq object, passes it throuh the `estimate_richness` function in `phyloseq` and returns the phyloseq object containing alpha diversity estimates as "alpha_[diversity_name]"
}
\examples{

data(GlobalPatterns)
phy_alpha_div_into_phy_metadata(GlobalPatterns,
     raref_depth = 5000,
     measures = c("Shannon", "Simpson"),
     verbose = TRUE)

}
