% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_pruneTaxa.R
\name{phy_pruneTaxa}
\alias{phy_pruneTaxa}
\title{Remove taxa that have counts summing to n}
\usage{
phy_pruneTaxa(physeq, n = 0)
}
\arguments{
\item{physeq}{A \code{phyloseq} object}

\item{n}{A \code{numeric} number of the pruning threshold to filter out any taxon that sums to strictly less than that value}
}
\value{
A \code{phyloseq} object with only taxa with \code{taxa_sums(physeq) > 0}
}
\description{
A simple wrapper function usable in a pipe that allows for quick pruning of a phyloseq object
}
\examples{

data(enterotype)

entero_noPyro <- subset_samples(enterotype, SeqTech != "Pyro454")

# before pruning
entero_noPyro

# after pruning
prune0taxa(entero_noPyro)

}
