% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_summarise_taxa_by_metadata.R
\name{phy_summarise_taxa_by_metadata}
\alias{phy_summarise_taxa_by_metadata}
\title{Summarise taxa by metadata}
\usage{
phy_summarise_taxa_by_metadata(physeq, metadata_var, .fun = mean)
}
\arguments{
\item{physeq}{A \code{phyloseq} object}

\item{metadata_var}{A \code{character} with the name of the factor to summarise by. it must be accepted by the code{group_by} dplyr function}

\item{.fun}{a \code{function} object to call by the summarise function in dplyr. I am thinking \code{mean, median, sum, scale, ...}}
}
\value{
A \code{matrix} object with the metadata variable in columns, and taxa in rows
}
\description{
A quick dplyr function that allows to summarise (mean, median, sum, ...) all taxa for each category of a metadata column. So far tested only on discrete variables.
}
\examples{

data(enterotype)
matrix_for_plot <- phy_summarise_taxa_by_metadata(microbiome::transform(enterotype, "clr"), metadata_var = "Enterotype", .fun = mean)
pheatmap::pheatmap(matrix_for_plot)

}
