% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_tax_glom.R
\name{phy_tax_glom}
\alias{phy_tax_glom}
\title{Agglomerate at higher taxonomic rank/level}
\usage{
phy_tax_glom(physeq, level, prefix = NULL, make.unique = TRUE)
}
\arguments{
\item{physeq}{A \code{phyloseq} object}

\item{level}{The taxonomic rank to aggregate to. Make sure it matches exactly the column name in the taxonomy table (see \code{rank_names(physeq)})}

\item{make.unique}{whether you should make unique the names. Default is \code{TRUE}. The standard separation pattern is \code{_.}}
}
\value{
An aggregated \code{phyloseq} object
}
\description{
Function that wraps speedyseq's tax_glom function and renames taxa based on more readable names as seen in the taxonomy table
}
\examples{

library(gautils2)
data("GlobalPatterns")
gp_Family <- phy_tax_glom(GlobalPatterns, "Family", make.unique = FALSE)
# this code shoots a "duplicated names" error. But you can make these names unique artificially:

gp_Family <- phy_tax_glom(GlobalPatterns, "Family", make.unique = TRUE)


}
