% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_transform.R
\name{phy_transform}
\alias{phy_transform}
\title{Transform microbiota counts}
\usage{
phy_transform(physeq, transform, binary_preval_thresh = 0)
}
\arguments{
\item{physeq}{a \code{phyloseq} object}

\item{transform}{\code{character} all options available in "transform()", extending it with the \code{arcsinh} ("asinh"), \code{geometric mean}("gm_mean"), \code{Hellinger} ("hellinger"), \code{binary}, and \code{Rank Inverse Normal transf.}("irn")}

\item{binary_preval_thresh}{\code{numeric}, the prevalence below which you want to transform your taxa into binary traits. default is \code{0}, which will not split the otu table}
}
\value{
A \code{phyloseq} object with sample-wise transformed data. also, it will create a phyloseq with Taxa in rows!
}
\description{
Same as transform(), but also adding further transformations, like hyperbolic arcsine, inverse-rank normal, binary transformation.
Additionally it can calculate geometric mean of a sample, although this is not really the same thing
}
\examples{

data(GlobalPatterns)
GP_Genus <- tax_glom(GlobalPatterns, "Genus")
tax <- "295395"
transf <- "clr"
transformed_GP <- transform_microbiota_ga(GP_Genus,transform = transf)


plot(density(abundances(GP_Genus)[tax,]), main = paste("taxon n°", tax), col = "blue")
lines(density(abundances(transformed_GP)[tax,]), col = "red")
legend("topright", legend = c("Untransformed", transf), col = c("blue", "red"), pch = 18)

}
