% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_to_fasta.R
\name{phyloseq_to_fasta}
\alias{phyloseq_to_fasta}
\title{Save Refseqs of phyloseq into a .fasta}
\usage{
phyloseq_to_fasta(physeq, destination_path = NULL, filename = NULL)
}
\arguments{
\item{physeq}{A \code{phyloseq} object}

\item{destination_path}{Write the file you want your fasta sequences to be written into. NB: the path must exist already}

\item{filename}{A \code{character}, the name of the file to write the files into. NB: the extension is not written automatically}
}
\value{
either a \code{character} with the sequences named after the taxon ID, or it writes them into the specified file in Fasta format
}
\description{
Write reference sequences of a phyloseq object to a text file with fasta formatting
}
