% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_to_picrust2.R
\name{phyloseq_to_picrust2}
\alias{phyloseq_to_picrust2}
\title{Phyloseq to Picrust input tables}
\usage{
phyloseq_to_picrust2(physeq, output.dir)
}
\arguments{
\item{physeq}{\code{phyloseq} object}

\item{output.dir}{The directory path to save your data into. NB: the function can create it}
}
\value{
a character vector saying where the files were stored
}
\description{
Phyloseq to Picrust input tables
}
\examples{

#There are not example datasets with also unique sequences, and I cannot provide our data due to our strict data protection policy

}
