% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLAR_results_kbl.R
\name{SOLAR_results_kbl}
\alias{SOLAR_results_kbl}
\title{Show clean SOLAR results}
\usage{
SOLAR_results_kbl(x, padj_method = "BH", complete.cases = TRUE)
}
\arguments{
\item{x}{A \code{data.frame} object as retrieved as-is by `get_SOLAR_results`}

\item{padj_method}{A \code{character} with the adjustment required. 'none' is also accepted and crude P-values will be kept}

\item{complete.cases}{A \code{logical} saying whether you want to remove rows with 1 or more NAs. Default is `TRUE`}
}
\value{
a \code{kbl} table, with default settings
}
\description{
Makes a kbl table from raw results obtained by by `get_SOLAR_results`, which is a useful function to retrieve multiple-trait heritability. But in principle it works even with just 1 trait
}
\examples{

# No microbiome SOLAR runs were stored as data.

}
