% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.pvalues.R
\name{table1.pvalues}
\alias{table1.pvalues}
\title{calculate p-values in table1 objects}
\usage{
table1.pvalues(x, ...)
}
\arguments{
\item{x}{a table in the `table1` set}

\item{...}{extra parameters. so far, none implemented, that I know}
}
\value{
a `table1` object
}
\description{
Inspired by the table1 vignette itself (https://cran.r-project.org/web/packages/table1/vignettes/table1-examples.html#example-a-column-of-p-values), only to have it quick every time I want a table1 significance
}
\examples{

data(mtcars)
table1::table1( ~ vs + gear + carb |as.factor(cyl), 
  data  = mtcars,
  extra.col = list("significance" = table1.pvalues),
  overall = "All cars"
  )
  
}
