% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLAR_results_kbl.R
\name{SOLAR_results_kbl}
\alias{SOLAR_results_kbl}
\title{Title}
\usage{
SOLAR_results_kbl(x, highlight_significant = 0.05)
}
\arguments{
\item{x}{An object obtained with \code{get_SOLAR_results}}

\item{highlight_significant}{\code{numeric} A heritability (h2) P-value threshold under which you would like to highlight the row}
}
\value{
a \code{kableExtra} object. ready to print as .html
}
\description{
Title
}
\examples{

tmp <- get_SOLAR_results(path_to_h2_runs)
print(SOLAR_results_kbl(tmp, 0.01))
}
